requirements.txt
Flask
gunicorn
flask-cors
pyodbc
python-dotenv
python-tds
